+++ fcat/searchOrtho.py
@@ -388,11 +388,12 @@ def searchOrtho(args):
     fcatFn.checkFileExist(annoDir, 'Please set path to annotation directory using --annoDir option.')
     for annoFile in glob.glob('%s/weight_dir/*.json' % args.coreDir):
         annoFileName = annoFile.split('/')[-1]
-        try:
-            os.symlink('%s/weight_dir/%s' % (args.coreDir, annoFileName), '%s/%s' % (annoDir, annoFileName))
-        except FileExistsError:
-            os.remove('%s/%s' % (annoDir, annoFileName))
-            os.symlink('%s/weight_dir/%s' % (args.coreDir, annoFileName), '%s/%s' % (annoDir, annoFileName))
+        if not os.path.exists('%s/%s' % (annoDir, annoFileName)):
+            try:
+                os.symlink('%s/weight_dir/%s' % (args.coreDir, annoFileName), '%s/%s' % (annoDir, annoFileName))
+            except FileExistsError:
+                os.remove('%s/%s' % (annoDir, annoFileName))
+                os.symlink('%s/weight_dir/%s' % (args.coreDir, annoFileName), '%s/%s' % (annoDir, annoFileName))
     annoQuery = args.annoQuery
 
     cpus = args.cpus
@@ -473,7 +474,7 @@ def searchOrtho(args):
     print('Done! Check output in %s' % fcatOut)
 
 def main():
-    version = '0.0.37'
+    version = '0.0.38'
     parser = argparse.ArgumentParser(description='You are running fcat version ' + str(version) + '.')
     required = parser.add_argument_group('required arguments')
     optional = parser.add_argument_group('optional arguments')
